fix: address CI lint failures and split use-onboarding-flow.ts
