+++ src/renderer/src/components/onboarding/use-onboarding-flow-types.ts
@@ -1,0 +1,13 @@
+export type StepNumber = 1 | 2 | 3 | 4
+export type StepId = 'agent' | 'theme' | 'notifications' | 'repo'
+
+export const STEPS: readonly {
+  id: StepId
+  stepNumber: StepNumber
+  valueKind: 'agent' | 'theme' | 'notifications' | 'repo'
+}[] = [
+  { id: 'agent', stepNumber: 1, valueKind: 'agent' },
+  { id: 'theme', stepNumber: 2, valueKind: 'theme' },
+  { id: 'notifications', stepNumber: 3, valueKind: 'notifications' },
+  { id: 'repo', stepNumber: 4, valueKind: 'repo' }
+]
